New regime for phase change
This code uses a simplified version of the equations for phase change used in CLM (see CLM documentation pg 110). The excess heat above freezing point (273.15K) is convrted into ecess energy, then used to melt a certain mass of ice. The new liquid water mass and the new ice mass are used to update the water fraction per layer, which can then percolate to the bottom of the snowpack. This is a modified version of the script snicar_driver_heating.m in the master branch.