new saving protocol
for each grain spec (i.e. radius and water coat thickness) produced by the grain evolution code, the driver now checks to see if the appropriate netcdf already exists in the wdir, and if so does not call the mie solver. If the file does not already exist, the mie solver is called, a new netcdf created and saved in the wdir in the correct filename format for CLM (4 digits front-padded with zeros)