Grain Size Routine ON/OFF
Now includes option to switch the grain size evolution routine on or off. This is achieved by defining the variable GRAIN_EVO as 1 (on) or 0 (off). Running the model multiple times with the grain evo on and then repeating with it off and differencing the mean of the two ensemble runs gives a measure of indirect albedo effect.