VERSION 0.1
Version 0.1 commit. 

Version 0.1 is controlled entirely from the CASPA_DRIVER script. The setup and CASPA scripts are formatted as functions to be called from the driver. The repository has been reorganised into folders that distinguish the V0.1 code from the development code. To download and run the impurity RIs and Mie codes need to be saved in the workspace too.