Incorporate dust as impurity, optionally update using scavenging rate
Code now includes dust i the upper layer as a uder defined variable. The dust concentration is held in variable 'y' and the initial mixing ratio at t=0 set by the user in the driver script. The dust then increases each timestep according to a scavenging rate, also defined by the user in the driver script. A scavenging rate of 2 means a doubling of the dust mixing ratio per timestep. Setting the scavenging ratio to 1 results in a constant dust mixing ratio throughout the model run. Setting initial dust to zero results in an entirely dust-free run.