Add mask function to block certain cells
This version includes a mask function that can be turned ON or OFF in the river. If ON, certain cells in the grid can be set to a constant spectral albedo that will remain constant throughout the run regardless of algal growth or grain evolution. Eventually this can be used to simulate outcrops of solid rock etc that interrupt the snowpack.

Currently the coordinates of the blocked areas have to be set manually inside CASPA_V01 - it would be good for a future commit to move the definition into the driver without breaking the script when MASK == 0. Also, it would be an improvement to load in a single 2D array instead of defining individual coordinates.