Significant Update: Now optionally masks using aerial images
New function (CASPA_mask_generator) called if MASK is switched on in the driver script. This reads in an aerial image of a snowpack, converts to greyscale, thesholds and converts to binary, giving an estimate of "snow" or "not snow" areas. Non-snow areas are then masked out of the CASPA cellular model. This currently uses test images saved in the workig directory - real snowpack imagery needed.